fix(ask-dev): break contracts_v2 import cycles for CodeQL (CHAOS-3294)
CodeQL reported 17 error-severity py/unsafe-cyclic-import alerts across
frame/embedded/validators, plus 5 py/unused-global-variable notes. All of the
cyclic alerts traced to one back-edge: validators imported frame for its
guard annotations while frame and embedded both imported validators to
register a no-answer policy at module scope, so a name could be read before
it was bound depending on which module an importer reached first.

Fixed by extracting the registration side into a leaf module. The policy
enum, per-model registry, canonical copy tables and the projection itself now
live in no_answer_policy, which imports nothing from this package except
base. embedded and frame register against that leaf directly, so the
registration graph no longer runs through validators. validators re-exports
every extracted name, keeping the guards' documented home and letting the
mutation tests patch one module object unchanged.

validators' remaining type-only reference to frame is now a module import
under TYPE_CHECKING rather than a name import, so no name is bound from a
partially-initialised module and the error-severity query no longer applies.
The runtime import graph is a DAG; verified by importing all thirteen modules
plus the package root in a fresh interpreter individually.

Unused globals resolved by reference, not suppression. ANSWERED_OUTCOMES,
EMPTY_CONTENT_OUTCOMES and IdempotencyKey are imported by sibling modules and
were simply missing from base's __all__. UNRESOLVED_OUTCOMES is now
load-bearing: DevResolutionEntry's else branch requires membership, so a new
ResolutionOutcome member cannot silently inherit "commits nothing" semantics.
contract_fixtures_v2's PREVIOUS_START was genuinely dead and is removed.

Verified with the CodeQL CLI locally against the same query suite CI runs
(security-and-quality): 18 cyclic and 6 unused-global alerts before, 0
error-severity after. No behaviour changes; 210 contract tests green.
