+++ src/dev_health_ops/api/dev/contracts_v2/base.py
@@ -31,10 +31,13 @@ from dev_health_ops.api.dev.contracts import (
 )
 
 __all__ = [
+    "ANSWERED_OUTCOMES",
+    "EMPTY_CONTENT_OUTCOMES",
     "Cardinality",
     "ContractModelV2",
     "EntityKind",
     "EvidenceHandle",
+    "IdempotencyKey",
     "Label",
     "LongText",
     "OpaqueID",
